re-enabled is_pod check in shared_array to break fixed_point. either memory allocation and destruction needs to be handled in shared_array specifically for fixed_point or, if not necessary, is_pod check can be modified to is_pod or is_fixed_point

+++ libs/math/include/math/shapeless_array.hpp
@@ -572,7 +572,7 @@ public:
     bool ret = true;
     for (std::size_t i = 0; ret && i < N; ++i)
     {
-      double va = this->At(i);
+      double va = static_cast<double>(this->At(i));
       if (ignoreNaN && std::isnan(va))
       {
         continue;
